\name{example_dat}
\alias{example_dat}
\docType{data}
\title{
Example Trawl and Stomach Contents Data
}
\description{
Example data containing trawl, called stomach, and full stomach contents data for consumption of sand lance by American plaice in NAFO Divisions 3LNO
}
\usage{data("example_dat")}
\format{
The format is:
\describe{
  \item{trawl}{Bottom trawl research survey data frame containing year: year that a survey occurred, and pa: presence (1)/absence (0) of sand lance in a trawl sample}
  \item{call_sto}{Called stomach samples from american plaice. Data frame containing year: year that a survey occurred, and pa: presence (1)/absence (0) of sand lance in a stomach sample}
  \item{full_sto}{Full stomach samples from american plaice. Data frame containing year: year that a survey occurred, and pa: presence (1)/absence (0) of sand lance in a stomach sample}
}
}
\source{
Data from DFO Newfoundland, raw data available upon request from DFO colleagues in Robertson et al. 2022 (https://doi.org/10.1111/2041-210X.13795)
}
\keyword{datasets}
