% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Create basic dataframe for the model}
\usage{
make_data(pa, year, n)
}
\arguments{
\item{pa}{Vector of presence/absence data from all surveys (binary)}

\item{year}{Vector of years associated with each observation (numeric)}

\item{n}{Vector of the number of observations for each survey (integer)}
}
\value{
dataframe
}
\description{
Combine presence/absence data from bottom trawl surveys, stomach contents, and an index for the different indices
}
\examples{
make_data(pa=rbinom(n=20, size=1, prob=0.5), year=rep(seq(from=1, to=10, by=1),2), n=c(10,10))
}
