% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_data}
\alias{model_data}
\title{Create data list for TMB}
\usage{
model_data(fishy_dat, k, chi, id, n)
}
\arguments{
\item{fishy_dat}{}

\item{k}{}

\item{chi}{}

\item{id}{}

\item{n}{}
}
\value{

}
\description{
Create data list for TMB
}
