% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_curve.R
\name{plot_curve}
\alias{plot_curve}
\title{Plot the functional response curve and associated data}
\usage{
plot_curve(rep, tmb_data)
}
\arguments{
\item{rep}{The report object from the TMB model}

\item{tmb_data}{The datalist used for the TMB model}
}
\value{
plot
}
\description{
Plot the functional response curve and associated data
}
\examples{
plot_curve(rep=rep, tmb_data=tmb_data)
}
