% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series of index and all dataseries}
\usage{
plot_ts(rep, sdrep, year, dat_names, ylim = c(0, 1))
}
\arguments{
\item{rep}{the report object from the TMB model}

\item{sdrep}{the variance report object from the TMB model}

\item{year}{a vector of the years used in the model}

\item{dat_names}{a vector of the names of the different data series used for the legend}

\item{ylim}{vector for y-limit in the plot (defaults to c(0,1))}
}
\value{

}
\description{
Plot time series of index and all dataseries
}
