% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sel_proc.R
\name{sel_proc}
\alias{sel_proc}
\title{Model selection procedure for real data}
\usage{
sel_proc(
  tmb_data,
  param_list,
  k_seq = seq(from = 0.1, to = 1, by = 0.05),
  chi_seq = seq(from = 0.05, to = 1, by = 0.05)
)
}
\arguments{
\item{tmb_data}{tmb data object from the model_data() function}

\item{param_list}{parameter list object from the model_param() function}

\item{k_seq}{seq of all possible k values to be cycled through in the model selection procedure, default is seq(from=0.1,to=1,by=0.05)}

\item{chi_seq}{seq of all possible chi values to be cycled through in the model selection procedure, default is seq(from=0.1,to=1,by=0.05)}
}
\value{

}
\description{
Model selection procedure for real data
}
