% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tastyfish-package.R
\docType{package}
\name{tastyfish-package}
\alias{tastyfish}
\alias{tastyfish-package}
\title{tastyfish: Account For Functional Response When Modeling Stomach Contents Data}
\description{
Using predator stomach contents data to model prey abundance is becoming increasingly common but can be biased if we 
  don't account for the functional response. This model will do both!
}
\keyword{internal}
