% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_data.R
\name{model_data}
\alias{model_data}
\title{Create data list for general functional response model}
\usage{
model_data(fishy_dat, n, type)
}
\arguments{
\item{fishy_dat}{Object output from the make_data() function}

\item{n}{A vector the length of the number of datasets used where each number indicates the number of observations for that dataset}

\item{type}{A character identifying whether you are creating data for a linear or nonlinear functional response model. Input "linear" for a linear model and "nonlinear" for a nonlinear model}
}
\value{
A list of data to be used when generating the TMB model. The list includes the total number of observations (n), the number of years (nyrs),
the number of datasets (ndex), an index vector for the year of each observation (iyear), an index vector for the dataset associated with each observation (idex), and
a vector for all presence/absence data (pa)
}
\description{
Create data list for general functional response model
}
\examples{
model_data(fishy_dat=fishy_dat, n=c(length(df$pa),length(df1$pa),length(df2$pa)), type="nonlinear")
}
