% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_param.R
\name{model_param}
\alias{model_param}
\title{Create parameter list for for general functional response model}
\usage{
model_param(tmb_data, lbeta)
}
\arguments{
\item{tmb_data}{TMB data object from the model_data() function}

\item{lbeta}{A logged starting value for the beta paramater}
}
\value{
A list of parameters for the functional response model, including a matrix with dimensions by number of years and datatypes(yr_tau_mat),
a vector the length of the number of years (iye), and a starting value for the beta parameter
}
\description{
Create parameter list for for general functional response model
}
\examples{
model_param(tmb_data=tmb_data, lbeta=log(5))
}
