% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_param.R
\name{model_param}
\alias{model_param}
\title{Create parameter list for TMB}
\usage{
model_param(tmb_data, lbeta)
}
\arguments{
\item{tmb_data}{}

\item{lbeta}{}
}
\value{

}
\description{
Create parameter list for TMB
}
