% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot time series of index and all dataseries}
\usage{
plot_ts(tmb_data, rep, sdrep, year, dat_names, ylim = c(0, 1), type)
}
\arguments{
\item{rep}{The report object from the TMB model}

\item{sdrep}{The variance report object from the TMB model}

\item{year}{A vector of the years used in the model}

\item{dat_names}{A vector of the names of the different data series used for the legend}

\item{ylim}{A vector for y-limit in the plot (defaults to c(0,1))}

\item{type}{A character identifying whether you are plotting a linear or nonlinear functional response model. Input "linear" for a linear model and "nonlinear" for a nonlinear model}
}
\value{
plot
}
\description{
Plot time series of index and all dataseries
}
\examples{
plot_ts(rep=rep, sdrep=sdrep, year=seq(from=1995, to=2018, by=1), dat_names=c("Trawl","Full Stomach Contents","Called Stomach Contents"), ylim=c(0.2,0.6))
}
